restructure and admin panel

+++ ch-11 Admin Panel/controller/emp.controller.js
@@ -14,16 +14,56 @@ const storage = multer.diskStorage({
 
 const upload = multer({storage});
 
-const homepage = async (req,res) => {
+const loginPage = (req,res) => {
+    return res.render("auth/login")
+}
+
+const checkAdmin = async (req,res) => {
+    console.log(req.body);
+
+    const admin = await Admin.findOne({email : req.body.email});
     
+    if (!admin && req.cookies.adminId == undefined) {
+        console.log("Admin not found");
+        return res.redirect("/")
+    }
+    
+    if (req.body.password != admin.password) {
+        console.log("password is incorrect");
+        return res.redirect("/")
+    }
+
+
+    res.cookie("adminId" , admin._id)
+    return res.redirect("/dashboard")
+}
+
+const logOut = (req,res) => {
+    res.clearCookie("adminId");
+    return res.redirect("/")
+}
+
+const homepage = async (req,res) => {
+    if (req.cookies.adminId == undefined) {
+        console.log("Admin not found");
+        return res.redirect("/")
+    }
     return res.render("dashboard")
 }
 
 const formPage = (req,res) => {
+    if (req.cookies.adminId == undefined) {
+        console.log("Admin not found");
+        return res.redirect("/")
+    }
     return res.render("form")
 }
 
 const tablePage = async (req,res) => {
+    if (req.cookies.adminId == undefined) {
+        console.log("Admin not found");
+        return res.redirect("/")
+    }
     const allAdmin = await Admin.find();
     return res.render("table",{allAdmin})
 }
@@ -35,11 +75,17 @@ const insertAdmin = async (req, res) => {
     }
 
     const newAdmin = await Admin.create(req.body);
-
+    console.log(req.cookies);
+    
+    res.cookie("name","Darshil Goyani");
     return res.redirect("/");
 };
 
 const deleteAdmin = async (req, res) => {
+    if (req.cookies.adminId == undefined) {
+        console.log("Admin not found");
+        return res.redirect("/")
+    }
     const deletedAdmin = await Admin.findByIdAndDelete(req.params.id);
 
     if (deletedAdmin) {
@@ -53,6 +99,10 @@ const deleteAdmin = async (req, res) => {
 };
 
 const updateAdminPage = async (req,res) => {
+    if (req.cookies.adminId == undefined) {
+        console.log("Admin not found");
+        return res.redirect("/")
+    }
     const updateAdmin = await Admin.findById(req.params.id); 
     return res.render("edit",{updateAdmin})
 }
@@ -77,4 +127,4 @@ const updateAdmin = async (req, res) => {
 
 
 
-module.exports = {homepage,formPage,tablePage,insertAdmin,upload,deleteAdmin,updateAdminPage,updateAdmin}
+module.exports = {homepage,formPage,tablePage,insertAdmin,upload,deleteAdmin,updateAdminPage,updateAdmin,loginPage,checkAdmin,logOut}
